rename fast_test_circuit, configure directories
