+++ pyproject.toml
@@ -33,3 +33,6 @@ zip-safe = false
 
 [tool.flake8]
 per-file-ignores = "__init__.py:F401"
+
+[tool.isort]
+skip = ["./quantestpy/__init__.py"]
