+++ quantestpy/__init__.py
@@ -1,2 +1,3 @@
-from .fast_test_circuit import FastTestCircuit
+from .simulator.pauli_circuit import PauliCircuit
 from .test_circuit import TestCircuit
+from .assertion.get_ctrl_val import assert_get_ctrl_val
