+++ test/simulator/test_pauli_circuit_assert_is_correct_qubit_val.py
@@ -1,17 +1,17 @@
 import random
 import unittest
 
-from quantestpy import FastTestCircuit
+from quantestpy import PauliCircuit
 from quantestpy.exceptions import QuantestPyTestCircuitError
 
 
-class TestFastTestCircuitAssertIsCorrectQubitVal(unittest.TestCase):
+class TestPauliCircuitAssertIsCorrectQubitVal(unittest.TestCase):
     """
     How to execute this test:
     $ pwd
     {Your directory where you git-cloned quantestpy}/quantestpy
     $ python -m unittest \
-        test.test_fast_test_circuit_assert_is_correct_qubit_val
+        test.simulator.test_pauli_circuit_assert_is_correct_qubit_val
     ....
     ----------------------------------------------------------------------
     Ran 4 tests in 0.002s
@@ -21,7 +21,7 @@ class TestFastTestCircuitAssertIsCorrectQubitVal(unittest.TestCase):
     """
 
     def test_regular(self,):
-        circ = FastTestCircuit(100)
+        circ = PauliCircuit(100)
         qubit_value = [random.randint(0, 1) for _ in range(60)]
         self.assertIsNone(
             circ._assert_is_correct_qubit_val(
@@ -31,7 +31,7 @@ class TestFastTestCircuitAssertIsCorrectQubitVal(unittest.TestCase):
         )
 
     def test_raise_from_qubit_val_incorrect_type(self,):
-        circ = FastTestCircuit(100)
+        circ = PauliCircuit(100)
         qubit_value = 0
         expected_error_msg = "test_qubit_value must be a list."
 
@@ -44,7 +44,7 @@ class TestFastTestCircuitAssertIsCorrectQubitVal(unittest.TestCase):
         self.assertEqual(cm.exception.args[0], expected_error_msg)
 
     def test_raise_from_val_incorrect_type(self,):
-        circ = FastTestCircuit(100)
+        circ = PauliCircuit(100)
         qubit_value = [0., 1]
         expected_error_msg = \
             "Values in test_qubit_value must be integer type."
@@ -58,7 +58,7 @@ class TestFastTestCircuitAssertIsCorrectQubitVal(unittest.TestCase):
         self.assertEqual(cm.exception.args[0], expected_error_msg)
 
     def test_raise_from_val_out_of_range(self,):
-        circ = FastTestCircuit(100)
+        circ = PauliCircuit(100)
         qubit_value = [0, 1, 2]
         expected_error_msg = \
             "Values in test_qubit_value must be either 0 or 1."
