+++ test/simulator/test_pauli_circuit_assert_is_correct_reg.py
@@ -1,16 +1,17 @@
 import random
 import unittest
 
-from quantestpy import FastTestCircuit
+from quantestpy import PauliCircuit
 from quantestpy.exceptions import QuantestPyTestCircuitError
 
 
-class TestFastTestCircuitAssertIsCorrectReg(unittest.TestCase):
+class TestPauliCircuitAssertIsCorrectReg(unittest.TestCase):
     """
     How to execute this test:
     $ pwd
     {Your directory where you git-cloned quantestpy}/quantestpy
-    $ python -m unittest test.test_fast_test_circuit_assert_is_correct_reg
+    $ python -m \
+        unittest test.simulator.test_pauli_circuit_assert_is_correct_reg
     ....
     ----------------------------------------------------------------------
     Ran 4 tests in 0.002s
@@ -20,14 +21,14 @@ class TestFastTestCircuitAssertIsCorrectReg(unittest.TestCase):
     """
 
     def test_regular(self,):
-        circ = FastTestCircuit(100)
+        circ = PauliCircuit(100)
         register = random.sample(range(0, 100), k=20)
         self.assertIsNone(
             circ._assert_is_correct_reg(reg=register, reg_name="test_register")
         )
 
     def test_raise_from_register_incorrect_type(self,):
-        circ = FastTestCircuit(100)
+        circ = PauliCircuit(100)
         register = {0, 1, 2}
         expected_error_msg = "test_register must be a list."
 
@@ -37,7 +38,7 @@ class TestFastTestCircuitAssertIsCorrectReg(unittest.TestCase):
         self.assertEqual(cm.exception.args[0], expected_error_msg)
 
     def test_raise_from_id_incorrect_type(self,):
-        circ = FastTestCircuit(100)
+        circ = PauliCircuit(100)
         register = [10.]
         expected_error_msg = "Indices in test_register must be integer type."
 
@@ -47,7 +48,7 @@ class TestFastTestCircuitAssertIsCorrectReg(unittest.TestCase):
         self.assertEqual(cm.exception.args[0], expected_error_msg)
 
     def test_raise_from_id_out_of_range(self,):
-        circ = FastTestCircuit(100)
+        circ = PauliCircuit(100)
         register = [99, 100]
         expected_error_msg = \
             "Qubit index 100 in test_register is out of range."
