+++ test/simulator/test_pauli_circuit_execute_all_gates.py
@@ -2,25 +2,25 @@ import unittest
 
 import numpy as np
 
-from quantestpy import FastTestCircuit
+from quantestpy import PauliCircuit
 
 
-class TestFastTestCircuitExecuteIthGate(unittest.TestCase):
+class TestPauliCircuitExecuteIAllGates(unittest.TestCase):
     """
     How to execute this test:
     $ pwd
     {Your directory where you git-cloned quantestpy}/quantestpy
-    $ python -m unittest test.test_fast_test_circuit_execute_i_th_gate
-    .....
+    $ python -m unittest test.simulator.test_pauli_circuit_execute_all_gates
+    .
     ----------------------------------------------------------------------
-    Ran 5 tests in 0.002s
+    Ran 1 test in 0.001s
 
     OK
     $
     """
 
     def setUp(self) -> None:
-        self.circ = FastTestCircuit(20)
+        self.circ = PauliCircuit(20)
         self.circ.add_gate(
             {"name": "x", "target_qubit": [0], "control_qubit": [1, 2],
              "control_value": [0, 0]})
@@ -40,28 +40,17 @@ class TestFastTestCircuitExecuteIthGate(unittest.TestCase):
     def tearDown(self) -> None:
         del self.circ
 
-    def test_x_gate(self,):
-        self.circ._execute_i_th_gate(i=0)
-        self.assertEqual(self.circ._qubit_value[0], 1)
+    def test_regular(self,):
+        self.circ._execute_all_gates()
+
+        self.assertEqual(self.circ._qubit_value[0], 0)
         self.assertEqual(self.circ._qubit_phase[0], 0.)
 
-    def test_y_gate(self,):
-        self.circ._execute_i_th_gate(i=1)
         self.assertEqual(self.circ._qubit_value[3], 1)
         self.assertEqual(self.circ._qubit_phase[3], np.pi/2.)
 
-    def test_z_gate(self,):
-        self.circ._execute_i_th_gate(i=2)
         self.assertEqual(self.circ._qubit_value[6], 0)
         self.assertEqual(self.circ._qubit_phase[6], 0.)
 
-    def test_swap_gate(self,):
-        self.circ._execute_i_th_gate(i=0)
-        self.circ._execute_i_th_gate(i=3)
         self.assertEqual(self.circ._qubit_value[0], 0)
         self.assertEqual(self.circ._qubit_value[10], 1)
-
-    def test_no_gate(self,):
-        self.circ._execute_i_th_gate(i=4)
-        self.assertEqual(self.circ._qubit_value[0], 0)
-        self.assertEqual(self.circ._qubit_phase[0], 0.)
